refactor(prerender): moved prerender middleware out to framework
