+++ apps/MFE/Footer/src/server/index.ts
@@ -2,11 +2,12 @@ import 'node-self'
 import path from 'path'
 
 // eslint-disable-next-line import/no-extraneous-dependencies
-import {helmetGuard, htmlMiddleware, httpLogger} from '@next/middlewares'
+import {helmetGuard, htmlMiddleware, httpLogger, prerenderMiddleware} from '@next/middlewares'
 import {json} from 'body-parser'
 import express from 'express'
 
-import {prerenderMiddleware} from './middleware/prerender'
+import federationStats from '../remote-entry/federation-stats.json'
+import remoteEntry from '../remote-entry/remote-entry'
 import {renderMiddleware} from './middleware/render'
 
 const publicPath = path.join(__dirname, '/public')
@@ -17,7 +18,7 @@ app.use(httpLogger(false))
 app.use(helmetGuard)
 
 app.use('/app', htmlMiddleware, renderMiddleware)
-app.use('/prerender', json(), prerenderMiddleware)
+app.use('/prerender', json(), prerenderMiddleware('mfe_footer', federationStats, remoteEntry))
 app.use('/', express.static(publicPath))
 
 export default app
