+++ libs/framework/middlewares/package.json
@@ -38,7 +38,9 @@
     "uuid": "^3.4.0"
   },
   "peerDependencies": {
-    "express": "^4.17.3"
+    "express": "^4.17.3",
+    "react": "18.0.0-beta-149b420f6-20211119",
+    "react-dom": "18.0.0-beta-149b420f6-20211119"
   },
   "gitHead": ""
 }
