+++ libs/framework/middlewares/src/index.ts
@@ -2,6 +2,7 @@ export * from './correlation'
 export * from './helmet'
 export * from './html'
 export * from './http-logger'
+export * from './prerender'
 export * from './site-url'
 export * from './cors'
 export * from './timeMachine'
